+++ js/challenges/word-problem.js
@@ -2,6 +2,7 @@ import { ChallengeBase } from './challenge-base.js';
 import * as sound from '../engine/sound.js';
 import { bigCorrectCelebration, splashEffect, goldenGlow } from '../engine/particles.js';
 import { pickCorrectPhrase } from '../engine/profile-theme.js';
+import { getCurrentProfileMeta, isJuniorTier } from '../state.js';
 
 export class WordProblem extends ChallengeBase {
     constructor(data, container) {
@@ -123,6 +124,10 @@ export class WordProblem extends ChallengeBase {
                     this.entry = '';
                     this._updateDisplay();
                     this.locked = false;
+                    if (isJuniorTier(getCurrentProfileMeta()) && this.attempts === 1 && this.hintsUsed === 0) {
+                        const hintBtn = this.container.querySelector('#hint-btn');
+                        if (hintBtn) hintBtn.click();
+                    }
                 }, 650);
             }
         });